refactor: 编辑器组件拆分
