+++ src/components/Editor/type.ts
@@ -1,0 +1,6 @@
+export interface SPL {
+  base64: string,
+  name: string,
+  id: string,
+  file: File,
+}
